% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{access_token}
\alias{access_token}
\title{Create a new Access Token}
\usage{
access_token(config, token_str)
}
\arguments{
\item{config}{An openid_config object}

\item{token_str}{A string containing the access token to decode}
}
\value{
An access_token object
}
\description{
Creates a new access token from a config and a token string
}
\keyword{internal}
