% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_trailing_slash}
\alias{add_trailing_slash}
\title{Add trailing slash to URL}
\usage{
add_trailing_slash(url)
}
\arguments{
\item{url}{A string containing a URL}
}
\value{
A string containing the URL with a trailing slash
}
\description{
If the app URL does not end with a slash, this function
will add one.
}
\keyword{internal}
