% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_cookie}
\alias{build_cookie}
\title{Build a cookie}
\usage{
build_cookie(key, value)
}
\arguments{
\item{key}{A string containing the cookie key}

\item{value}{A string containing the cookie value}
}
\value{
A string containing the cookie
}
\description{
Builds an HttpOnly cookie from a key and value
}
\keyword{internal}
