% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_redirect_uri}
\alias{build_redirect_uri}
\title{Build a redirect URI}
\usage{
build_redirect_uri(app_url)
}
\arguments{
\item{app_url}{A string containing the app URL with a trailing slash}
}
\value{
A string containing the redirect URI
}
\description{
Builds a redirect URI from an app URL
}
\keyword{internal}
