% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_do_nothing.R
\name{call_do_nothing}
\alias{call_do_nothing}
\title{Return the input}
\usage{
call_do_nothing(data)
}
\arguments{
\item{data}{Anything.}
}
\value{
The inpput
}
\description{
Return the input
}
\examples{
call_do_nothing("a")
}
