% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{decode_token}
\alias{decode_token}
\title{Decode a token}
\usage{
decode_token(config, token)
}
\arguments{
\item{config}{An openid_config object}

\item{token}{The token to decode}
}
\value{
A list containing the decoded token's data
}
\description{
Decodes a token
}
\keyword{internal}
