% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{expires_at}
\alias{expires_at}
\title{Get the expiration date and time of an access token}
\usage{
expires_at(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A POSIXct object containing the date and time the token expires
}
\description{
Gets the expiration date and time of an access token
}
