% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{expires_in}
\alias{expires_in}
\title{Get the expiration time of an access token}
\usage{
expires_in(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A duration object containing the time until the token expires
}
\description{
Gets the expiration time of an access token
}
