% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_access_token}
\alias{get_access_token}
\title{Get the access token string}
\usage{
get_access_token(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A string containing the access token
}
\description{
Gets the access token string
}
\keyword{internal}
