% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_bearer}
\alias{get_bearer}
\title{Get the Authorization header for an access token}
\usage{
get_bearer(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A string containing the Authorization header
}
\description{
Gets the Authorization header for an access token
}
\keyword{internal}
