% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_client_id}
\alias{get_client_id}
\title{Get the client ID}
\usage{
get_client_id(config)
}
\arguments{
\item{config}{An openid_config object}
}
\value{
A string containing the client ID
}
\description{
Gets the client ID for the provider
}
\keyword{internal}
