% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_login_url}
\alias{get_login_url}
\title{Get the login URL for the app}
\usage{
get_login_url(config)
}
\arguments{
\item{config}{An openid_config object}
}
\value{
A string containing the login URL
}
\description{
Gets the URL that the provider should redirect to
after a login attempt.
}
\keyword{internal}
