% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_logout_url}
\alias{get_logout_url}
\title{Get the logout URL for the provider}
\usage{
get_logout_url(config)
}
\arguments{
\item{config}{An openid_config object}
}
\value{
A string containing the logout URL
}
\description{
Gets the URL for SLO (single logout)
}
\keyword{internal}
