% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token_field}
\alias{get_token_field}
\title{Get the issued at time of an access token}
\usage{
get_token_field(token, field)
}
\arguments{
\item{token}{An access_token object}

\item{field}{The field to get from the token}
}
\value{
A POSIXct object containing the date and time the token was issued
}
\description{
Gets the issued at time of an access token
}
