% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{is_expired}
\alias{is_expired}
\title{Check if an access token is expired}
\usage{
is_expired(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A boolean indicating if the token is expired
}
\description{
Checks if an access token is expired
}
