% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{is_valid}
\alias{is_valid}
\title{Check if an access token is valid}
\usage{
is_valid(token)
}
\arguments{
\item{token}{An access_token object}
}
\value{
A boolean indicating if the token is valid
}
\description{
Checks if an access token is valid
by checking if it is expired
}
