% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth0.R
\name{new_auth0_config}
\alias{new_auth0_config}
\title{Create a new auth0_config object}
\usage{
new_auth0_config(client_id, client_secret, auth0_domain, app_url)
}
\arguments{
\item{client_id}{The client ID for the app}

\item{client_secret}{The client secret for the app}

\item{auth0_domain}{The domain for the Auth0 tenant}

\item{app_url}{The URL for the app}
}
\value{
An auth0_config object
}
\description{
Creates a new auth0_config object
}
