% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entra_id.R
\name{new_entra_id_config}
\alias{new_entra_id_config}
\title{Create a new entra_id_config object}
\usage{
new_entra_id_config(tenant_id, client_id, client_secret, app_url)
}
\arguments{
\item{tenant_id}{The tenant ID for the app}

\item{client_id}{The client ID for the app}

\item{client_secret}{The client secret for the app}

\item{app_url}{The URL for the app}
}
\value{
An entra_id_config object
}
\description{
Creates a new entra_id_config object
}
