% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entra_id.R
\name{new_entra_id_config_from_env}
\alias{new_entra_id_config_from_env}
\title{Create a new entra_id_config object from environment variables}
\usage{
new_entra_id_config_from_env()
}
\value{
An entra_id_config object
}
\description{
Creates a new entra_id_config object using environment variables:
\itemize{
\item TAPLOCK_ENTRA_ID_CLIENT_ID
\item TAPLOCK_ENTRA_ID_CLIENT_SECRET
\item TAPLOCK_ENTRA_ID_TENANT_ID
\item TAPLOCK_APP_URL
\item TAPLOCK_ENTRA_ID_USE_REFRESH_TOKEN (Optional)
}
}
