% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{new_google_config_from_env}
\alias{new_google_config_from_env}
\title{Create a new google_config object from environment variables}
\usage{
new_google_config_from_env()
}
\value{
A google_config object
}
\description{
Creates a new google_config object using environment variables:
\itemize{
\item TAPLOCK_GOOGLE_CLIENT_ID
\item TAPLOCK_GOOGLE_CLIENT_SECRET
\item TAPLOCK_APP_URL
\item TAPLOCK_GOOGLE_USE_REFRESH_TOKEN (Optional)
}
}
