% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keycloak.R
\name{new_keycloak_config}
\alias{new_keycloak_config}
\title{Create a new keycloak_config object}
\usage{
new_keycloak_config(
  base_url,
  realm,
  client_id,
  client_secret,
  app_url,
  use_refresh_token = TRUE,
  validate_audience = TRUE
)
}
\arguments{
\item{base_url}{The base URL for the Keycloak instance}

\item{realm}{The realm for the app}

\item{client_id}{The client ID for the app}

\item{client_secret}{The client secret for the app}

\item{app_url}{The URL for the app}

\item{use_refresh_token}{Enable the use of refresh tokens}

\item{validate_audience}{Whether to validate the token audience against the client ID}
}
\value{
A keycloak_config object
}
\description{
Creates a new keycloak_config object
}
