% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_openid_config}
\alias{new_openid_config}
\title{New openid configuration}
\usage{
new_openid_config(provider, app_url, ...)
}
\arguments{
\item{provider}{The openid provider to use}

\item{app_url}{The URL of the application
(used to build redirect, login, and logout URLs)}

\item{...}{Additional arguments passed to the provider's configuration.
This depends on the provider.

The \code{"google"} provider accepts the following arguments:
\itemize{
\item \code{client_id}
\item \code{client_secret}
}

The \code{"entra_id"} provider accepts the following arguments:
\itemize{
\item \code{client_id}
\item \code{client_secret}
\item \code{tenant_id}
}

The \code{"auth0"} provider accepts the following arguments:
\itemize{
\item \code{client_id}
\item \code{client_secret}
\item \code{auth0_domain}
}}
}
\value{
An openid_config object
}
\description{
Creates a new openid configuration object
for the given provider. You can use this function or
the individual provider functions.
}
