% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_cookies}
\alias{parse_cookies}
\title{Parse cookies}
\usage{
parse_cookies(x)
}
\arguments{
\item{x}{A string containing the cookies}
}
\value{
A list containing the cookies
}
\description{
Parses cookies from a string
}
\keyword{internal}
