% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_bearer}
\alias{remove_bearer}
\title{Remove bearer from token}
\usage{
remove_bearer(token)
}
\arguments{
\item{token}{A string containing the token}
}
\value{
A string containing the token without the "Bearer " prefix
}
\description{
Removes the "Bearer " prefix from a token
}
\keyword{internal}
