% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{request_token}
\alias{request_token}
\title{Request a token from the provider}
\usage{
request_token(config, authorization_code)
}
\arguments{
\item{config}{An openid_config object}

\item{authorization_code}{The authorization code to use}
}
\value{
An access_token object
}
\description{
Requests a token from the provider
}
\keyword{internal}
