% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sso_shiny_app}
\alias{sso_shiny_app}
\title{Create a Shiny app with SSO}
\usage{
sso_shiny_app(config, ui, server)
}
\arguments{
\item{config}{An openid_config object}

\item{ui}{A Shiny UI function}

\item{server}{A Shiny server function. This function requires
all three arguments:  \code{input}, \code{output}, and \code{session}.}
}
\value{
A Shiny app (Compatible with \code{\link[shiny:shinyApp]{shinyApp()}})
}
\description{
Creates a Shiny app with SSO (single sign-on)
based on the given configuration.
}
\seealso{
\code{\link[=new_openid_config]{new_openid_config()}}
}
