% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapLock-package.R
\docType{package}
\name{tapLock-package}
\alias{tapLock}
\alias{tapLock-package}
\title{tapLock: Seamless Single Sign-On}
\description{
Swift and seamless Single Sign-On (SSO) integration. Designed for effortless compatibility with popular Single Sign-On providers like Google and Microsoft, it streamlines authentication, enhancing both user experience and application security. Elevate your R-based projects for a simplified, unified, and secure authentication process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/tapLock}
  \item \url{https://maurolepore.github.io/tapLock/}
  \item Report bugs at \url{https://github.com/maurolepore/tapLock/issues}
}

}
\author{
\strong{Maintainer}: FIXME \email{maurolepore@gmail.com}

}
\keyword{internal}
