% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{token}
\alias{token}
\title{Get the access token}
\usage{
token(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A Shiny session}
}
\value{
An access_token object
}
\description{
Gets the access token from the session to be used
for internal logic.
}
