% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{use_futures}
\alias{use_futures}
\title{Use futures for asynchronous computations}
\usage{
use_futures(plan = future::multicore, workers = 1)
}
\arguments{
\item{plan}{A plan object. Defaults to a multicore plan.}

\item{workers}{Number of workers to use. Defaults to 1.}
}
\value{
This function is called for its side effect.
}
\description{
Enable a future plan for asynchronous computations.
Since tapLock needs to do calls to external APIs, it can be a good idea
to use future to make the calls asynchronous.

This function is just meant as a convenience function for the user.
We recommend that you read the documentation for the future package
to understand how to use it.
}
