% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{add_auth_layers}
\alias{add_auth_layers}
\title{Add authentication middle ware to a 'tower' object}
\usage{
add_auth_layers(tower, config)
}
\arguments{
\item{tower}{A 'tower' object from the package 'tower'}

\item{config}{An 'openid_config' object}
}
\value{
A modified 'tower' object with authentication layers
}
\description{
Attaches the necessary authentication layers
to a 'tower' object. This will secure any layer added
after.
}
