% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{new_google_config}
\alias{new_google_config}
\title{Create a new google_config object}
\usage{
new_google_config(client_id, client_secret, app_url)
}
\arguments{
\item{client_id}{The client ID for the app}

\item{client_secret}{The client secret for the app}

\item{app_url}{The URL for the app}
}
\value{
A google_config object
}
\description{
Creates a new google_config object
}
