% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keycloak.R
\name{new_keycloak_config_from_env}
\alias{new_keycloak_config_from_env}
\title{Create a new keycloak_config object from environment variables}
\usage{
new_keycloak_config_from_env()
}
\value{
A keycloak_config object
}
\description{
Creates a new keycloak_config object using environment variables:
\itemize{
\item TAPLOCK_KEYCLOAK_URL
\item TAPLOCK_KEYCLOAK_REALM
\item TAPLOCK_KEYCLOAK_CLIENT_ID
\item TAPLOCK_KEYCLOAK_CLIENT_SECRET
\item TAPLOCK_APP_URL
\item TAPLOCK_KEYCLOAK_USE_REFRESH_TOKEN (Optional)
}
}
