% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{print.access_token}
\alias{print.access_token}
\title{Print an access token}
\usage{
\method{print}{access_token}(x, ...)
}
\arguments{
\item{x}{An access_token object}

\item{...}{Ignored}
}
\value{
No return value, called for side effects
}
\description{
Prints an access token's expiration date
}
