% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapLock-package.R
\docType{package}
\name{tapLock-package}
\alias{tapLock}
\alias{tapLock-package}
\title{tapLock: Seamless Single Sign-On for R Applications}
\description{
Swift and seamless Single Sign-On (SSO) integration. Designed for effortless compatibility with popular 'SSO' providers like Google and Microsoft, tapLock streamlines authentication, enhancing both user experience and application security. Elevate your R-based projects with tapLock for a simplified, unified, and secure authentication process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/tapLock}
  \item \url{https://maurolepore.github.io/tapLock/}
  \item Report bugs at \url{https://github.com/maurolepore/tapLock/issues}
}

}
\author{
\strong{Maintainer}: ixpantia, SRL \email{hola@ixpantia.com}

}
\keyword{internal}
